Move histograms from buckets to allocators
This moves the mark and available histograms outside of the Bucket type
and into the local and mailbox allocators, reducing the memory required
per process by at least 100 bytes. Behaviour wise nothing changes, as
for the LocalAllocator we would evacuate objects for all Buckets the
moment we'd find one that was fragmented.