Stop storing garbage collection counts
Storing the garbage collection counts per process is a waste of space,
as most processes will never read this data. Removing this saves us 8
bytes per process. In the future we may use synchronised global counters
that store collection counts per process, and are updated by the garbage
collector. For now, we won't store anything as we don't expose the data
yet anyway.