Removed storing of entire process statuses
Obtaining the process status has always been a bit questionable. For
one, it's not particularly useful to see that a process is running or
being garbage collected. Second, it requires a full 8 bytes of memory
per process to store.

In this commit, we drop the storing of full process statuses, and add a
boolean flag "waiting for message" that we use instead where necessary.
Currently this won't reduce the size of a process due to alignment
requirements, but in the future we may be able to work around this by
reducing the size of other fields.