Move all "class" methods to module methods
In the past I have gone back and forth a bit on the idea of using
class/static methods or not. With this commit I'm making a final
decision on this topic: Inko will not have class/static methods, at
least not until 1.0. Instead, Inko will use module methods. This means
that the following methods have been changed:

* HashMap.from_array  -> std::hash_map.from_array
* Trait.implement_for -> std::trait.implement
* Integer.from_string -> std::integer.parse
* Float.from_string   -> std::float.parse

To make this happen some compiler changes had to be made to optimise
`hash_map.from_array` and to use `trait.implement` instead of
`Trait.implement_for`.