Fix storing module methods in globals
Due to the use of an incorrect scope object, the compiler would use -1
as the index for methods stored in module globals. Due to overflowing
this would result in the VM allocating memory for 65 535 globals,
resulting in at least 512 KB of memory being used for every module that
defined a module method. Worse, if a module were to use these methods
the VM might end up executing the wrong method due to all of them using
the same slot.