Refactor the global allocator and block sizes
This refactors the global allocator so that it doesn't require on
explicit locking as much, and changes various parts of Immix blocks (and
tests) to no longer rely directly on the block size. This makes it a bit
easier to adjust the block size in the future, without having to fix
various tests.