Fix non-blocking socket reads
Prior to this commit, if a read size was given we would use Rust's Take
trait to read up to N bytes from a socket. Unfortunately, sockets don't
produce an EOF until they are closed, which meant the read would
continue indefinitely. To work around this we use the same buffering
logic as is used by our recv_from implementation.