Fix rehashing of HashMap
The in-place rehashing approach was broken. Depending on the hash value
of a pair, it could be placed in the wrong bucket. This commit fixes the
problem by using a new Array for the buckets. This is less memory
efficient, but:

1. Correct
2. Simpler

Making the rehashing logic more memory efficient is something we may
look into in the future.