Require all attributes to be assigned in "init"
This adds compiler validation to ensure that all object attributes are
assigned a value in the "init" method. This validation step only looks
at top-level assignments in the "init" method. This is the only way to
reliably check this at this time, as no control-flow analysis is
available in the compiler. This is also the easiest to implement, and
prevents somebody from writing code like this:

    def init {
      some_condition.if_true {
        @number = 10
      }
    }

Here the attribute might be left unassigned if `if_true` does not run.
Requiring assignments at the top would require one to write something
like this:

    def init {
      @number = some_condition.if true: {
        ...
      }, false: {
        ...
      }
    }