Implement Inko's parser in Inko
The parser is an LL(1) recursive descent parser. While the Ruby parser
was used as a reference, it is not a 1:1 port.

As part of porting the parser to Inko, several syntax changes were made:

1. "where" on methods is not supported by the Inko parser. When we start
   using the parser we will move "where" to traits.

2. The comments "#!" and "##" are replaced with just "#". This makes the
   lexer and parser internals easier, and removes the need for having to
   remember three different comment types.

3. Mutable rest arguments are supported in the Inko parser using
   "mut *NAME", instead of "* mut NAME".

Using the parser is straightforward:

    import std::compiler::parser::Parser

    let parser = Parser.new(input: '10 + 2', file: 'test.inko')

    try! parser.parse

The AST nodes are located in modules under `std::compiler::ast`. For
example, integer literals are in `std::compiler::ast::literals`. For now
there is some minor duplication across some of the AST nodes. For
example, the types of methods (MethodDefinition) and required methods
(RequiredMethodDefinition) are similar. When we start implementing the
compiler we may change this around a bit, or perhaps split the AST nodes
into more separate types.

Fixes https://gitlab.com/inko-lang/inko/issues/172