Do not expose Trait and Module by default
This allows custom modules to define types with the name Trait or
Module. To use Trait and Module you just import them from std::trait and
std::module respectively. The module std::module does not define
anything useful at this time, but more useful code will be added to it
in the future.