Add Iterator.partition and std::pair
The method Iterator.partition can be used to partition an Iterator into
a Pair of Array objects. The return type is the newly introduced
std::pair::Pair, which is a simple binary tuple. We also introduce
std::pair::Triple for ternary pairs. We do not introduce any further
tuple types, as custom objects are better suited for pairs of more than
three values.