Remove async finalisation of objects
Finalising objects was handled in one of two ways:

1. When reclaiming blocks we would schedule blocks that needed
   finalisation. These blocks were processed in a separate thread.
2. When allocating into a block that needed finalisation, all objects in
   this block in need of finalisation would be finalised.

The idea of this approach was to move the expensive finalising of
objects out of the garbage collection phase, at the cost of delaying
finalisation a bit.

Unfortunately, this approach was anything but simple. This finalisation
approach also required various data structures taking up memory.
Finally, as part of the finalisation process we would have to scan over
all objects; instead of just lines.

In this commit we change the approach to a much simpler one. All async
related code and data structures are removed. When allocating an Immix
block, we zero out the block.

When allocating a new object into a previously used slot that needs
finalising, we finalise the object. This delays the finalising of
objects until their memory is reused. In the usual program this should
not pose a problem, as memory will be reused frequently. Indeed, this is
something we have observed with Inko's own test suite and a few other
test programs: memory usage actually goes down due to fewer data
structures needed, instead of going on.