Use bit flags for tracking a process' status
Using the bits in a single byte allows us to track different flags of a
process (e.g. the blocking or main flag) in a single byte. This also
allows us to track the termination status of a process, without needing
extra memory.

The use of this new status field also allows us to stop sending messages
to a process that has terminated.

This fixes https://gitlab.com/inko-lang/inko/issues/186