Lower collection thresholds from 8/16 to 2/4
When Immix block sizes were reduced from 32 KB to 8 KB, the amount of
objects needed to trigger a collection increased. The young generation
was using a threshold of 8 MB, and the mature generation a threshold of
16 MB. At 32 KB per block this translates to roughly one collection
every 256 new Immix blocks, while at 8 KB a collection would happen
every 1024 new Immix blocks.

Reducing the young threshold to 2 MB and the mature threshold to 4 MB
means we once again collect after 256 new Immix blocks in the young
generation, and 512 new Immix blocks in the mature generation.