Add ByteArray to the prelude
The compiler needs support for circular types (e.g. A depending on B and
B depending on A). The way ByteArray was set up in std::byte_array meant
this could not be done with multiple passes, as the assignment of the
`ByteArray` constant would require full type-inference too early in the
compilation process. Moving ByteArray to the prelude and setting it up
while bootstrapping simplifies this, and makes it easier for the user to
use byte arrays.