Refactor std::mirror
This is a pretty big refactor of std::mirror to make it entirely
optional. As part of this std::format has been refactored as well, and
all implementations of the Inspect trait (now called Format) have been
moved into std::mirror.

Formatting objects is now part of std::mirror as that's where it
belongs, since the output is often based on the internal representation
of an object. The new API for pretty printing objects is a bit more
verbose, but not that much:

    import std::mirror

    10.mirror.inspect # => '10'

All of this means that the only module currently using std::module is
std::test::assert, which needs std::mirror to produce useful test
failures. Other code that previously used "inspect" has been refactored
to no longer depend on it.