Add Path.directory
This method can be used to obtain a Path pointing to the directory of
another Path, without touching the file system. A simple example:

    import std::fs::path::Path

    Path.new('/foo/bar/baz').directory.to_string # => '/foo/bar'

The internals of Path.directory depend on the newly introduced
std::fs::path::bits module. This module will be used for defining
internal methods used as the building blocks of the Path type. This
allows us to offload work to separate methods (so we can return early
where needed), without exposing all this in the public API of
std::fs::path.