Clean up Boolean using an "if" instruction
By introducing the virtual "if" instruction we can clean up Boolean so
that we no longer have to refine the True and False instances. If we
also clean up Nil in a similar way it becomes easier to rewrite the
compiler in Inko, as it no longer needs to support the refining of
random objects.