Revert Boolean and Nil cleanup
I thought this would make the self-hosting compiler easier to implement,
but it turns out this may not be the case. If we use two objects per
type (one for the "class" and one for the instance), then yes it will
clean things up. But this two object approach complicates the compiler
in various other places. Balancing the two, I prefer being able to
refine random objects as this means fewer VM changes are needed for for
example Boolean and Nil.