Remove support for extending traits
Now that the Ruby compiler defines types using multiple passes, there no
longer is a need for forward declaring and extending traits. Extending
traits could also lead to odd behaviour when adding default methods. For
example:

    trait A {}

    object Foo {}

    impl A for Foo {}

    trait A {
      def thing -> Integer {
        10
      }
    }

Here type "Foo" would not have access to method "thing", because it is
added after "A" is implemented for "Foo".

Removing support for this entirely solves this problem, and we no longer
need it anyway.