Remove support for constant receivers
This removes support for syntax such as `A::B` where `A` and `B` are
constants. Inko has not allowed defining nested types or constants in
types for a while, and this syntax was only used in a few places to
bootstrap the runtime. Removing support for this syntax simplifies both
the parser and compiler quite a bit.