Don't use shared Windows runners
With the release of shared Windows runners on GitLab.com, using runners
with _just_ the "windows" tag results in CI using these shared runners.
The shared runners don't have the right software installed, and are
still in beta. To prevent using these, we now also require the runners
to specify the "inko" tag.